Pass an empty body for enable/disable recurring payments
We were getting an error from the API saying that we weren't
providing the correct Content-Type. However, we were generating
the request from the SDK so we stepped through how the request
was being created on the SDK-side of things and found that the
headers were being stripped from within Axios:
    https://github.com/axios/axios/blob/649d739288c8e2c55829ac60e2345a0f3439c730/dist/axios.js#L1449-L1450

After some discussion, the correct solution would be to update
the API to not require a Content-Type header whenever there is
no body since it's not necessary, and is a waste of bytes over
the wire.

The current patch to the SDK is a temporary fix until we remove
the restriction from the API.

Full context: https://highbeamco.slack.com/archives/C02JB5M1599/p1660766477125359